check user del Pro, up Bio, up trans

+++ controllers/product.controller.js
@@ -374,6 +374,16 @@ module.exports = {
     }
 
     try {
+      // Check if Delete user is product owner
+      const userProduct = await Product.findOne({
+        where: { id: req.params.id },
+      });
+      if (userProduct.seller_id !== req.user.id) {
+        return res.status(401).json({
+          type: "UNAUTHORIZED",
+          message: "Unauthorized Access",
+        });
+      }
       // Delete product pictures
       await deleteProductImages(req.params.id);
